# Improve debug performance of update_with_buffer on windows