Hello,
I think replacing `buffer.iter().cloned().collect()` with `buffer.to_vec()` can noticeably improve the time spent in `update_with_buffer()` on Windows - Debug builds. 
When compiling with `--release` there is not much of a performance difference.
Here are the times spent in `window.update_with_buffer(&buffer)` using a `1280x720` framebuffer.

|   | Before | After |
|---|---|---|
| Debug | `120 ms` | `52 ms` |
| Release | `2.4 ms` | `2.2 ms` |